adding more integration tests to books

+++ server/api/book/controller.ts
@@ -4,6 +4,11 @@ import {Request, Response} from "express";
 import Util from "../utils";
 
 export default class BookController {
+    /**
+     * Creates a book and sets the current user as owner
+     * @param req 
+     * @param res 
+     */
     public static create(req:Request, res:Response):void {
         var myBook = new Book({
             name: req.body.name,
@@ -22,13 +27,23 @@ export default class BookController {
         .catch(Util.handleError(res));
     }
 
+    /**
+     * Updates a book, only the owner or editors can make updates
+     * @param req
+     * @param res 
+     */
     public static update(req:Request, res:Response):void {
         var myId:string = req.body._id;
         var myBook = new Book(req.body);
         delete req.body._id;
+
         myBook.validate()
         .then(pass => Book.findById(myId))
         .then(oldBook => {
+            if (oldBook.owner !== req.user._id && oldBook.editors.indexOf(req.user._id) === -1) {
+                throw Error(`User: ${req.user._id} is neither an owner or editor of Book: ${myId}`)
+            }
+
             return Book.findByIdAndUpdate(myId, req.body).exec()
             .then(data => BookSocket.onAddOrChange(myBook, oldBook))
         })
@@ -36,15 +51,23 @@ export default class BookController {
         .catch(Util.handleError(res));
     }
 
+    /**
+     * Deletes a book, only the owner can do this action
+     * @param req 
+     * @param res 
+     */
     public static remove(req:Request, res:Response):void {
         var myId:string = req.params.id;
 
         Book.findById(myId).exec()
         .then(book => {
+            if (book.owner !== req.user._id) {
+                throw `User ${req.user._id} is not owner and cannot delete book: ${book._id}`
+            }
             return book.remove()
             .then(() => BookSocket.onDelete(book))
         })
-        .then(Util.handleResponseNoData(res))
+        .then(() => res.json().status(200))
         .catch(Util.handleError(res));
     }
